feat(serving): implement deployment simulation and monitoring infrastructure (phase 5)
Introduces the full Phase 5 infrastructure: a FastAPI scoring service,
containerized deployment, transaction simulation, PSI-based drift monitoring,
performance tracking, and automated retraining simulation.

Scoring API (app/main.py):
- FastAPI service with POST /predict (batch scoring) and GET /health endpoints
- Loads model artifact and metadata (feature_list, threshold) on startup via
  lifespan context manager
- Stores runtime state in app.state; raises typed HTTPException (503/422)
- Enforces exact feature contract from training metadata; fails fast on mismatch
- Structured logging via Python logging module

Containerization:
- Dockerfile based on python:3.10-slim-bookworm; copies only artifacts/models/
- .dockerignore excludes raw data, monitoring outputs, notebooks, and tests
- Production image uses requirements.txt only (no dev/test dependencies)

Transaction Simulation (scripts/simulate_transactions.py):
- Loads IEEE-CIS dataset, shuffles, and sends configurable batches to the API
- JSON sanitization: NaN/inf replaced with None before serialization
- Graceful batch-level error handling; failed batches are logged and skipped
- Produces predictions CSV with timestamp, TransactionID, fraud_probability,
  fraud_flag, isFraud, and TransactionAmt

Drift Monitoring (src/utils/drift.py, scripts/monitor_model.py):
- PSI implementation with quantile-based binning derived from reference
- Zero-bin proportions replaced with epsilon (1e-4) instead of being skipped,
  preventing silent underestimation of drift
- Handles edge cases: constant reference, all-zero current counts (extreme drift)
- monitor_model.py validates input paths, computes PSI per feature, and saves
  timestamped drift and performance reports to artifacts/monitoring/

Retraining Simulation (scripts/retrain_model.py):
- Reads drift report JSON, computes max PSI across features
- Triggers run_training_pipeline only when max_psi exceeds configured threshold
- Logs decision and new artifact paths

Dependencies:
- Split requirements.txt (production) from requirements-dev.txt (test + notebooks)
- Removed annotated-doc (unrecognized package); removed Jupyter and pytest from
  production image dependencies

Bug fixes:
- scripts/train_model.py: added PROJECT_ROOT sys.path injection for consistent
  src imports when called outside the project root
- src/data/schema.py: corrected TransactionDT expected dtype from float64 to int64

Tests:
- tests/test_api_scoring.py: fixture-based TestClient with monkeypatched lifespan
  to avoid loading real artifacts in CI; covers happy path, batch scoring,
  high-risk model, missing feature (422), and model not loaded (503)
- tests/test_drift_metrics.py: six cases covering identical distributions,
  significant drift (asserts PSI > 0.2), skewed distributions, constant reference,
  small arrays, and non-overlapping distributions (asserts finite result)

Documentation:
- docs/09_deployment_and_monitoring.md: added H1 title, Mermaid architecture
  diagram, Quick Start section with step-by-step commands, and corrected
  inaccurate description of the inference feature pipeline

Closes #9

+++ src/data/schema.py
@@ -19,7 +19,9 @@ NO_MISSING_COLUMNS = {
 
 EXPECTED_DTYPES: Dict[str, str] = {
     "TransactionID": "int64",
-    "TransactionDT": "float64",
+    # In the raw dataset TransactionDT is typically an integer-like time index.
+    # We only require it to be stored as 64-bit integer here.
+    "TransactionDT": "int64",
     "TransactionAmt": "float64",
     "isFraud": "int64",
 }
